requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==3.1.6
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.12
